Update README.md
The code is not running out of the box. The changes in application.properties(under resources), the additional comments in readme.md and an additional step in the console is required to make it run out of the box. I have tested it on Windows 10(with java 8), Ubutnu Linux 18.04 LTS and MacOS Sierra(Java 8).