These changes are required to make it work
These changes are required to make it work with 5.1.1 Couchbase server. 
I am assuming that a user "travel-sample" will be created with the admin access involving access to the "travel-sample" otherwise the default Out of the box solution is not going to work. For simplicity i have put the password as "password" for almost anything(As this is not the production instance)